fix: refactor inner classes and long strings
